tudo ok

+++ CartasSuperTrunfo.c
@@ -8,8 +8,8 @@ int main() {
     float area1 = 331.4;                                                // Área em km²
     float pib1 = 105.0;                                                 // PIB em bilhões
     int pontosTuristicos1 = 15;                                         // Número de pontos turísticos
-    double densidade1 = populacao1 / area1;                             // Densidade populacional    
-    double pibPerCapita1 = (pib1 * 1000000000) / populacao1;            // PIB per capita em reais
+    float densidade1 = populacao1 / area1;                              // Densidade populacional
+    float pibPerCapita1 = (pib1 * 1000000000) / populacao1;             // PIB per capita em reais
 
 
     // Carta 2 - Fortaleza
@@ -19,8 +19,9 @@ int main() {
     float area2 = 314.9;                                                // Área em km²
     float pib2 = 67.0;                                                  // PIB em bilhões
     int pontosTuristicos2 = 20;                                         // Número de pontos turísticos
-    double densidade2 = populacao2 / area2;                             // Densidade populacional
-    double pibPerCapita2 = (pib2 * 1000000000) / populacao2;            // PIB per capita em reais
+    float densidade2 = populacao2 / area2;                              // Densidade populacional
+    float pibPerCapita2 = (pib2 * 1000000000) / populacao2;             // PIB per capita em reais
+
 
 
     // calculo do super poder //
